Changing default surpressException to false
The reason why i think this should be false, is whenever something in the server breaks, the one handling the errors should be dependant on the user, and not ignored.

Since for example, whenever there is downtime on jikan, and you get an exception you know that the server has an exception.
But it with being default on, and not knowing there was an option to disable that supression, i would have returned incorrect information back to the user.

With it being default on, just returning a null could mean anything. "User not found" or "Server is dead" and there is a much higher chance of these errors go unnoticed, since it would just show as if that request just does not exist.

Thats why i think its much better for it to be optional to suppress warnings, like how you would do in any systems that have exceptions or warnings.